Decouple canvas and sketch and refactor to separate Canvas components
