+++ src/app/page.tsx
@@ -1,7 +1,6 @@
-import MotionTitle from '@/components/MotionTitle'
-import P5Canvas from '@/components/P5Canvas'
-
+import BoidCanvas from '@/components/BoidCanvas'
 import Contact from '@/components/Contact'
+import MotionTitle from '@/components/MotionTitle'
 import Projects from '../components/Projects'
 export const revalidate = false // SSG at build; set to a number for ISR
 
@@ -11,7 +10,7 @@ export default function Home() {
       <header className='relative overflow-hidden grid items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]'>
         <MotionTitle />
         <div className='absolute inset-0 -z-10'>
-          <P5Canvas />
+          <BoidCanvas />
         </div>
       </header>
       <main>
